feat: add kdoc
