+++ instantsearch-compose/src/main/java/com/algolia/instantsearch/compose/paging/SearcherPager.kt
@@ -3,9 +3,18 @@ package com.algolia.instantsearch.compose.paging
 import androidx.paging.PagingData
 import kotlinx.coroutines.flow.Flow
 
+/**
+ * Pager for Search.
+ */
 public interface SearcherPager<T : Any> {
 
+    /**
+     * A cold Flow of PagingData, which emits new instances of PagingData once they become invalidated.
+     */
     public val flow: Flow<PagingData<T>>
 
+    /**
+     * Notify searcher's configuration has changed.
+     */
     public fun notifySearcherChanged()
 }
