+++ instantsearch-compose/src/main/java/com/algolia/instantsearch/compose/searchbox/internal/SearchBoxComponents.kt
@@ -1,0 +1,57 @@
+package com.algolia.instantsearch.compose.searchbox.internal
+
+import androidx.compose.animation.*
+import androidx.compose.animation.core.LinearOutSlowInEasing
+import androidx.compose.animation.core.tween
+import androidx.compose.foundation.clickable
+import androidx.compose.material.Icon
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.TextFieldColors
+import androidx.compose.material.TextFieldDefaults
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Close
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+
+/**
+ * Search Box default colors.
+ */
+@Composable
+internal fun SearchColors(): TextFieldColors {
+    return TextFieldDefaults.textFieldColors(
+        backgroundColor = MaterialTheme.colors.background,
+        focusedIndicatorColor = Color.Transparent,
+        unfocusedIndicatorColor = Color.Transparent,
+        disabledIndicatorColor = Color.Transparent,
+    )
+}
+
+/**
+ * Search box default clear icon.
+ */
+@OptIn(ExperimentalAnimationApi::class)
+@Composable
+internal fun SearchClearIcon(
+    visible: Boolean,
+    onClick: () -> Unit
+) {
+    AnimatedVisibility(
+        visible = visible,
+        enter = slideInHorizontally(
+            initialOffsetX = { fullWidth -> fullWidth / 3 },
+            animationSpec = tween(100, easing = LinearOutSlowInEasing)
+        ) + fadeIn(),
+        exit = slideOutHorizontally(
+            targetOffsetX = { fullWidth -> fullWidth / 3 },
+            animationSpec = tween(100, easing = LinearOutSlowInEasing)
+        ) + fadeOut()
+    ) {
+        Icon(
+            imageVector = Icons.Default.Close,
+            contentDescription = null,
+            tint = MaterialTheme.colors.onBackground,
+            modifier = Modifier.clickable(onClick = onClick)
+        )
+    }
+}
