+++ instantsearch-compose/src/main/java/com/algolia/instantsearch/compose/stats/internal/StatsComposeImpl.kt
@@ -3,6 +3,9 @@ package com.algolia.instantsearch.compose.stats.internal
 import androidx.compose.runtime.MutableState
 import com.algolia.instantsearch.compose.stats.StatsCompose
 
+/**
+ * [StatsCompose] implementation.
+ */
 internal class StatsComposeImpl<T>(
     override val stats: MutableState<T>
 ) : StatsCompose<T> {
